Textures a rajoute

+++ src/utils.c
@@ -6,13 +6,13 @@
 /*   By: facarval <facarval@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/10 14:08:55 by facarval          #+#    #+#             */
-/*   Updated: 2024/04/12 11:30:25 by facarval         ###   ########.fr       */
+/*   Updated: 2024/04/15 09:55:33 by facarval         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../cub3d.h"
 
-int	ft_abs(int nb)
+double	ft_abs(double nb)
 {
 	if (nb < 0)
 		return (-nb);
